Move display column factory
